+++ apps/dashboard/components/landing-page/HeroChatGPTPlugin.jsx
@@ -253,7 +253,7 @@ export function HeroChatGPTPlugin() {
         </div>
 
         <div className="flex justify-center mt-10 gap-x-6">
-          {/* <Button target="_blank" href="https://app.chaindesk.ai/signin">Start Building</Button> */}
+          {/* <Button target="_blank" href="https://development.chatsappai.com/signin">Start Building</Button> */}
           {/* <Button
           href="https://www.youtube.com/watch?v=dQw4w9WgXcQ"
           variant="outline"