+++ apps/lp/app/(default)/integrations/page.tsx
@@ -66,7 +66,7 @@ export default function IntegrationsPage({
         description={`Train and Integrate your AI Agent with your favorite tools and platforms: WhatsApp, Shopify, WordPress, and more.`}
         cta={{
           label: 'Get Started for Free',
-          url: 'https://app.chaindesk.ai/agents',
+          url: 'https://development.chatsappai.com/agents',
         }}
         // youtubeVideoId={product?.youtubeVideoId}
         // imageUrl={product?.imageUrl}